reorganize edit profile function locations
